implement user trading

+++ src/main/java/com/minhle/cryptotrading/crypto_trading_system/repository/AggregatedPriceRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface AggregatedPriceRepository extends JpaRepository<AggregatedPrice, Integer> {
+public interface AggregatedPriceRepository extends JpaRepository<AggregatedPrice, Long> {
     Optional<AggregatedPrice> findTopBySymbolOrderByCreatedAtDesc(String symbol);
 }
